My_researchers/requirements.txt
mcp
python-dotenv
requests
uvicorn
fastapi
notebooklm-py
